Move workio into host addon
